Rework method to create SB page structure.
Add functions splitall() and setup_subparents().

setup_subparents() improves establishing the SB page structure by
basing the process on the list of xml files. This fixes issue #29, only
folders containing XMLs eventually are created.

Add wait and check to delete_all_children(), which allows subpages to
be deleted and the rest of the page creation and population to be
executed in the same call. fixes issue #36

Fixed upload_files_matching_xml() to not try to upload folders if they
match.

Improved some print statements for clarity.